Directory reorganisation. Split the various proofs into their own sub-folders for clarity
