+++ packages/aztec.js/test/abiEncoder/joinSplit/joinSplit.spec.js
@@ -1,12 +1,12 @@
 const chai = require('chai');
 const { padLeft } = require('web3-utils');
 
-const aztecProof = require('../../src/proof/joinSplit');
-const secp256k1 = require('../../src/secp256k1');
-const sign = require('../../src/sign');
-const note = require('../../src/note');
-const abiEncoder = require('../../src/abiEncoder/joinSplit');
-const { K_MAX } = require('../../src/params');
+const aztecProof = require('../../../src/proof/joinSplit');
+const secp256k1 = require('../../../src/secp256k1');
+const sign = require('../../../src/sign');
+const note = require('../../../src/note');
+const abiEncoder = require('../../../src/abiEncoder/joinSplit');
+const { K_MAX } = require('../../../src/params');
 
 const { expect } = chai;
