+++ packages/aztec.js/test/abiEncoder/joinSplit/outputCoder.spec.js
@@ -1,11 +1,11 @@
 const chai = require('chai');
 const { padLeft } = require('web3-utils');
 
-const bn128 = require('../../src/bn128');
-const secp256k1 = require('../../src/secp256k1');
-const note = require('../../src/note');
-const outputCoder = require('../../src/abiEncoder/outputCoder');
-const { K_MAX } = require('../../src/params');
+const bn128 = require('../../../src/bn128');
+const secp256k1 = require('../../../src/secp256k1');
+const note = require('../../../src/note');
+const outputCoder = require('../../../src/abiEncoder/joinSplit');
+const { K_MAX } = require('../../../src/params');
 
 const { expect } = chai;
